+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Unreleased
 
-## Features
+## New Features
 * routing - a template argument was added to the `@routing.route` decorator.
   This argument determines which templates a route can be added to.
   https://github.com/anvilistas/anvil-extras/issues/293
@@ -8,7 +8,7 @@
   https://github.com/anvilistas/anvil-extras/pull/298
 * routing - `@routing.redirect()` decorator added
   https://github.com/anvilistas/anvil-extras/pull/298
-
+* hashlib module added
 
 
 # v2.0.1 16-Mar-2022